+++ CHANGELOG.md
@@ -14,12 +14,10 @@ Revisions:
 Features:
 
   - Direct PDBID search
-  > Custom GLmol menu for WebGL devices
+  - Custom GLmol menu for WebGL devices
   > Custom protein coloring in GLmol menu
   > Custom chain representation in GLmol menu
   > Biological assembly in GLmol menu
-  > Query parameters for GLmol representation
-  > Query parameters for biological assembly
 
 ## 2.1.3 (2014-07-07)